Rework segment writing to avoid race/improve perf
There was race when writing to segments under high concurrency
due to the ringbuf being fully utilized and buffers being
reused before fully flushed to disk.

To fix the race, we remove the ringbuf as straight mutexes
with buffered writer is performing better and is simpler to
reason about.

+++ pkg/wal/reader.go
@@ -52,6 +52,6 @@ func (s *SegmentReader) Close() error {
 	return s.f.Close()
 }
 
-func (s *SegmentReader) SampleMetadata() (t SampleType, sampleCount uint16) {
+func (s *SegmentReader) SampleMetadata() (t SampleType, sampleCount uint32) {
 	return s.iter.Metadata()
 }
